+++ examples/tests/pose_accuracy_coco_val.sh
@@ -2,24 +2,15 @@
 
 
 
+# USAGE EXAMPLE
+# clear && clear && make all -j`nproc` && bash ./examples/tests/pose_accuracy_coco_val.sh
+
 # Script for internal use. We might completely change it continuously and we will not answer questions about it.
 
 clear && clear
 
-# USAGE EXAMPLE
-# clear && clear && make all -j`nproc` && bash ./examples/tests/pose_accuracy_coco_test.sh
-
-# # Go back to main folder
-# cd ../../
-
-
-# Write COCO-format JSON
-# Last id:
-    # ID 20671  <-->    #frames = 1471      -->     ~ 1.5 min at 15fps
-    # ID 50006  <-->    #frames = 3559      -->     ~ 4 min at 15fps
-
 # Parameters
-IMAGE_FOLDER=/home/gines/devel/images/val2017/
+IMAGE_FOLDER=~/devel/images/val2017/
 JSON_FOLDER=../evaluation/coco_val_jsons/
 OP_BIN=./build/examples/openpose/openpose.bin